% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_resolution.R
\name{ag_res_output}
\alias{ag_res_output}
\title{Write csv files of Agency/Program Standard Taxonomic Resolution}
\usage{
ag_res_output(long, bioregion)
}
\arguments{
\item{long}{= long data frame containing taxonomic counts.}

\item{col.ord}{= a list of names used to sort the final output.}

\item{taxon}{= The taxon of interest.}
}
\value{
Writes multiple csv files containing the percentage of taxa
identified by each agency/program at the order, family and genus
level.
}
\description{
Write csv files of Agency/Program Standard Taxonomic Resolution
}
