% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_resolution.R
\name{agency_count2}
\alias{agency_count2}
\title{Agency/Program count and percentage per bioregion}
\usage{
agency_count2(long)
}
\arguments{
\item{long}{= long data frame containing taxonomic counts.}

\item{bioregion}{= The bioregion code used to aggregate the data.}
}
\value{
The number and percentage of samples that each agency/bioregion
contributes to a bioregion of interest.
}
\description{
Agency/Program count and percentage per bioregion
}
