% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_resolution.R
\name{agency_resolution}
\alias{agency_resolution}
\title{Test Agency/Program Standard Taxonomic Resolution}
\usage{
agency_resolution(long, col.ord, rank)
}
\arguments{
\item{long}{= long data frame containing taxonomic counts.}

\item{col.ord}{= a list of names used to sort the final output.}

\item{taxon}{= The taxon of interest.}
}
\value{
Returns the percentage of taxa identified by each agency/program
at the specified taxonomic rank.
}
\description{
Test Agency/Program Standard Taxonomic Resolution
}
