% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{agency_taxa_diff}
\alias{agency_taxa_diff}
\title{Check for Differences in the Taxa Identified by Agency}
\usage{
agency_taxa_diff(Long, Level)
}
\arguments{
\item{Long}{= Long data frame, typically representing a single bioregion.}

\item{Level}{= The taxonomic rank to perform the comparison.}
}
\value{
Compares the taxa identified by each agency.
}
\description{
Check for Differences in the Taxa Identified by Agency
}
