% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{agg_catchment_map}
\alias{agg_catchment_map}
\title{Aggregate scores by catchment}
\usage{
agg_catchment_map(map.station, map.me, all.data, pct_sum, spatial, taxon.rank,
  todays.date, agg.col = "FEATUREID", catch.count = 3, sub.dir)
}
\description{
Aggregate scores by catchment
}
