% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{agg_huc12_map}
\alias{agg_huc12_map}
\title{Aggregate scores by HUC12}
\usage{
agg_huc12_map(map.station, map.me, pct_sum, spatial, taxon.rank, todays.date,
  agg.col = "STATION_ID", catch.count)
}
\description{
Aggregate scores by HUC12
}
