% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{all_tables}
\alias{all_tables}
\title{All Tables}
\usage{
all_tables(spatial, calc.date, m.c, all.data, tab.event,
  todays_date = format(Sys.time(), "\%m_\%d_\%y"), random_event, main.dir,
  previous.subdir, map.agg.col)
}
\arguments{
\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\description{
All Tables
}
