% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{auto_bio}
\alias{auto_bio}
\title{Automate Spatial Classification}
\usage{
auto_bio(my.data, index_res, bioregion = NULL)
}
\arguments{
\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\description{
Automate Spatial Classification
}
