% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{auto_index}
\alias{auto_index}
\title{Combine Multiple Automation Functions to Create an Index}
\usage{
auto_index(my.data, bioregion, taxon_rank, calc.date = "11_09_16",
  jack.runs = 100, todays_date = format(Sys.time(), "\%m_\%d_\%y"),
  seed = TRUE)
}
\arguments{
\item{bioregion}{=}

\item{todays_date}{=}

\item{raw.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{taxon.rank}{=}
}
\description{
Combine Multiple Automation Functions to Create an Index
}
