% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{auto_metrics}
\alias{auto_metrics}
\title{Automate Metric Calculation}
\usage{
auto_metrics(all.data, index.res, bioregion, taxon.rank,
  todays_date = format(Sys.time(), "\%m_\%d_\%y"))
}
\arguments{
\item{all.data}{=}

\item{bioregion}{=}

\item{taxon.rank}{=}

\item{todays_date}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}
}
\description{
Automate Metric Calculation
}
