% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{auto_plot}
\alias{auto_plot}
\title{Automate Plots}
\usage{
auto_plot(plot.score, plot.value, taxon_rank, bioregion,
  todays_date = format(Sys.time(), "\%m_\%d_\%y"))
}
\arguments{
\item{bioregion}{=}

\item{todays_date}{=}

\item{raw.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{taxon.rank}{=}
}
\description{
Automate Plots
}
