% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{auto_score}
\alias{auto_score}
\title{Automate Scoring}
\usage{
auto_score(raw.data, bioregion, taxon_rank, prod_date, metric.freq,
  master.df = master, metric.class = m.c, todays_date = format(Sys.time(),
  "\%m_\%d_\%y"), num_itr = "10i", previous.sub.dir)
}
\arguments{
\item{raw.data}{=}

\item{bioregion}{=}

\item{todays_date}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{taxon.rank}{=}
}
\description{
Automate Scoring
}
