% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{bibi_2011_vs_2016}
\alias{bibi_2011_vs_2016}
\title{2011 Ratings vs. 2016 Ratings}
\usage{
bibi_2011_vs_2016(curr.dir, spatial, taxon.rank, todays.date)
}
\arguments{
\item{curr.dir}{= the directory to the 2016 data.}

\item{todays.date}{= todays date.}
}
\value{
Provides two tables: 1) Coastal rating comparisons and 
2) Inland rating comparisons.
}
\description{
2011 Ratings vs. 2016 Ratings
}
