% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{bioregion_agg}
\alias{bioregion_agg}
\title{Bioregion aggregation}
\usage{
bioregion_agg(Bioregion.df)
}
\arguments{
\item{Bioregion.df}{= Bioregion data frame to be aggregated}
}
\value{
Prepares the bioregion data for metric calculations.
}
\description{
Bioregion aggregation
}
