% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{category_counts}
\alias{category_counts}
\title{Category Counts}
\usage{
category_counts(spatial, taxon.rank, todays.date)
}
\arguments{
\item{spatial}{= "BASIN", "REGION", or "BIOREGION".}

\item{taxon.rank}{= "ORDER", "FAMILY", or "GENUS".}

\item{todays.date}{= todays date.}
}
\value{
Provides a table of Category counts (Ref, Min, Mix, Mod, Deg) by
bioregion.
}
\description{
Category Counts
}
