% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpss.R
\name{cf_score}
\alias{cf_score}
\title{Calculate the metrics from GLIMPSS}
\usage{
cf_score(metrics, disturbance, sc_ceiling, sc_floor)
}
\arguments{
\item{metrics}{= a data frame containing metric values}

\item{disturbance}{= does the metric increase or decrease with disturbance? 
Requires: "INCREASE" or "DECREASE".}

\item{sc_ceiling}{= the upper threshold typically defined by the
5th or 95th percentile.}

\item{sc_floor}{= the lower threshold typically defined by the
5th or 95th percentile.}
}
\value{
A gradient scoring approach.
}
\description{
Calculate the metrics from GLIMPSS
}
