% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{clean_char}
\alias{clean_char}
\title{Clean Up Character Vectors}
\usage{
clean_char(x)
}
\arguments{
\item{x}{= A vector.}
}
\value{
Cleans up a character vector by verifying that it is class character,
all characters are uppercase, and all leading/trailing white space has been
removed.
}
\description{
Clean Up Character Vectors
}
