% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{clean_taxa}
\alias{clean_taxa}
\title{Clean Taxa}
\usage{
clean_taxa(taxa.long)
}
\arguments{
\item{taxa.long}{= a data frame, in a long data format, containing taxonomic
hierarchy columns.}
}
\value{
Standardizes taxa according to the Chessie BIBI protocol.
}
\description{
Clean Taxa
}
