% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_resolution.R
\name{common_agency_discrepancies2}
\alias{common_agency_discrepancies2}
\title{Checks for common Agency/Program Discrepancies}
\usage{
common_agency_discrepancies2(long)
}
\arguments{
\item{long}{= long data frame containing taxonomic counts.}

\item{bioregion}{= The bioregion code used to aggregate the data.}
}
\value{
Checks for common agency/program discrepancies at the order, family,
and genus level.
}
\description{
Checks for common Agency/Program Discrepancies
}
