% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subdir.R
\name{create_subdir}
\alias{create_subdir}
\title{Create a Subdirectory for BIBI Outpt}
\usage{
create_subdir(main.dir, spatial.res, taxon.res)
}
\arguments{
\item{main.dir}{<- Specificy the main directory (file path).}

\item{spatial.res}{<- Specify the spatial resolution of the data.}

\item{taxon.res}{<- Specify the taxonomic resolution of the data.}
}
\description{
Create a Subdirectory for BIBI Outpt
}
