% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpss.R
\name{glimpss_metrics}
\alias{glimpss_metrics}
\title{Calculate the metrics from GLIMPSS}
\usage{
glimpss_metrics(Long, index)
}
\arguments{
\item{Long}{= a long data frame of taxonomic data}

\item{index}{= specify the regional and season index. 
Requires: Mountain Spring ("MT_SP"), Mountain Summer ("MT_SU"), 
Plateau Spring ("PL_SP"), or Plateau Summer ("PL_SU").}
}
\value{
Calculates the metrics from the Genus Level Index of most 
Probable Stream Status (GLIMPSS) (POND et al. 2013).
}
\description{
Calculate the metrics from GLIMPSS
}
