% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{group.me.low.res}
\alias{group.me.low.res}
\title{Group Me Low Resolution}
\usage{
group.me.low.res(front, proc_date, data_set, job)
}
\arguments{
\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\description{
Group Me Low Resolution
}
