% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{group.me2}
\alias{group.me2}
\title{Group Me 2}
\usage{
group.me2(spatial, taxon_rank, data_set, calc.date, final_score = FALSE)
}
\arguments{
\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\description{
Group Me 2
}
