% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_resolution.R
\name{group.names}
\alias{group.names}
\title{Data frame of Taxa of Interest}
\usage{
group.names(long, low.rank, upper.rank, taxon)
}
\arguments{
\item{long}{= long data frame containing taxonomic counts.}

\item{low.rank}{= The lower of taxonomic rank of interest.}

\item{upper.rank}{= The upper of taxonomic rank of interest; corresponds with
the taxon of interest.}

\item{taxon}{= The taxon of interest.}
}
\value{
Returns a data frame at the taxonomic level specified by low.rank,
containing only lower level taxa of the taxon of interest.
}
\description{
Data frame of Taxa of Interest
}
