% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{identify_duplicates}
\alias{identify_duplicates}
\title{Identify Duplicate Samples}
\usage{
identify_duplicates(long.df)
}
\arguments{
\item{long.df}{= Data in a long data format.}
}
\value{
Identify duplicate samples and suggest how the samples should be
handled.
}
\description{
Identify Duplicate Samples
}
