% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{import_msa_data}
\alias{import_msa_data}
\title{Import MS Access Data Sheets}
\usage{
import_msa_data(odbc.name)
}
\arguments{
\item{odbc.name}{= A character string that refers to the odbc connection
name.}
}
\value{
Assigns Chessie BIBI related ecoregion and bioregion distinctions.
}
\description{
Import MS Access Data Sheets
}
