% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_and_whisker_plots.R
\name{index_plot2}
\alias{index_plot2}
\title{Index Plot}
\usage{
index_plot2(index.df, condition.colname, condition.order = NULL,
  index.colname = "INDEX", threshold, plot.title)
}
\description{
Index Plot
}
