% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{index_table}
\alias{index_table}
\title{Index Table}
\usage{
index_table(spatial, taxon_rank, data_set, calc.date,
  todays_date = format(Sys.time(), "\%m_\%d_\%y"))
}
\arguments{
\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\description{
Index Table
}
