% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_dev.R
\name{iterate_dev}
\alias{iterate_dev}
\title{Iterative Index Development}
\usage{
iterate_dev(all.data, index.res, taxon.rank, runs = 10,
  todays_date = format(Sys.time(), "\%m_\%d_\%y"), unique.station = FALSE,
  standard_count = FALSE,
  main.dir = "D:/ZSmith/Projects/Chessie_BIBI/Output")
}
\arguments{
\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\description{
Iterative Index Development
}
