% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{jack_table}
\alias{jack_table}
\title{Jackknife Table}
\usage{
jack_table(spatial, data_set, calc.date, todays_date = format(Sys.time(),
  "\%m_\%d_\%y"), maindir = main.dir, prevdir = previous.subdir)
}
\arguments{
\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\description{
Jackknife Table
}
