% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macs.R
\name{mac}
\alias{mac}
\title{Middle Atlantic Coastal Plain}
\usage{
mac(Info, Long, taxa.rank, tol_col, habit_col)
}
\arguments{
\item{Info}{= Taxonomic Information}

\item{Long}{= Taxonomic data in long format}

\item{tol_col}{= Specify Tolerance value column.}

\item{habit_col}{= Specify Habit column.}
}
\value{
Prepare the taxonomic data and score all metrics for each sites
within the Middle Atlantic Coastal Plain bioregion
}
\description{
Middle Atlantic Coastal Plain
}
