% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macs.R
\name{mac_clean_taxa}
\alias{mac_clean_taxa}
\title{MAC: Standarize Taxonomic Resolution}
\usage{
mac_clean_taxa(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Removes and adjusts taxonomic resolution to make data sets from
multiple sources comparable.
}
\description{
MAC: Standarize Taxonomic Resolution
}
