% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_this}
\alias{map_this}
\title{Map the BIBI Ratings}
\usage{
map_this(spatial, calc.date, taxon.rank, all.data, tab.events, todays.date,
  title.me, very.poor.thresh, random_event, sub.dir, agg.col)
}
\arguments{
\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\value{
Multiple tables related to aggregated scores and ratings, as wel as,
a .png image of the map representing the HUC12 ratings.
}
\description{
Map the BIBI Ratings
}
