% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macs.R
\name{metrics_mac}
\alias{metrics_mac}
\title{Middle Atlantic Coastal Plain Metrics}
\usage{
metrics_mac(master, Long, taxa.rank, tol_col, habit_col)
}
\arguments{
\item{tol_col}{= Specify Tolerance value column.}

\item{habit_col}{= Specify Habit column.}

\item{Info}{= Taxonomic Information}

\item{Family}{= Familial taxonomic counts}

\item{Order}{= Ordinal taxonomic counts}

\item{Long_Data}{= Taxonomic data in long format}

\item{Level}{= Taxonomic level ("FAMILY" or "GENUS")}
}
\value{
2011 Chessie BIBI metrics for the northern appalachian plateau (Middle Atlantic Coastal Plain) ecoregion
}
\description{
Middle Atlantic Coastal Plain Metrics
}
