% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napu.R
\name{metrics_napu}
\alias{metrics_napu}
\title{NAPU Metrics}
\usage{
metrics_napu(master, Long, taxa.rank, tol_col, ffg_col)
}
\arguments{
\item{Long}{= Taxonomic data in long format}

\item{tol_col}{= Specify Tolerance value column.}

\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{Info}{= Taxonomic Information}
}
\value{
2011 Chessie BIBI metrics for the northern appalachian plateau
(NAPU) ecoregion.
}
\description{
NAPU Metrics
}
