% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.R
\name{metrics_nca}
\alias{metrics_nca}
\title{NCA metrics}
\usage{
metrics_nca(master, Long, tol_col, ffg_col)
}
\arguments{
\item{Long}{= Taxonomic data in long format}

\item{tol_col}{= Specify Tolerance value column.}

\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{Info}{= Taxonomic Information}
}
\value{
2011 Chessie BIBI metrics for the north central Appalachians (NCA) ecoregion
}
\description{
NCA metrics
}
