% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piedmont.R
\name{metrics_pied}
\alias{metrics_pied}
\title{Piedmont metrics}
\usage{
metrics_pied(master, Long, taxa.rank, tol_col, ffg_col)
}
\arguments{
\item{Long}{= Taxonomic data in long format}

\item{tol_col}{= Specify Tolerance value column.}

\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{Info}{= Taxonomic Information}
}
\value{
2011 Chessie BIBI metrics for the Piedmont ecoregion
}
\description{
Piedmont metrics
}
