% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridges.R
\name{metrics_ridges}
\alias{metrics_ridges}
\title{Ridges metrics}
\usage{
metrics_ridges(master, Long, ffg_col, habit_col)
}
\arguments{
\item{Long}{= Taxonomic data in long format}

\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{habit_col}{= Specify Habit column.}

\item{Info}{= Taxonomic Information}
}
\value{
2011 Chessie BIBI metrics for the ridges ecoregion
}
\description{
Ridges metrics
}
