% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep.R
\name{metrics_sep}
\alias{metrics_sep}
\title{Southeastern Plains Metrics}
\usage{
metrics_sep(master, Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}

\item{Info}{= Taxonomic Information}
}
\value{
2011 Chessie BIBI metrics for the northern appalachian plateau
(Southeastern Plains) ecoregion
}
\description{
Southeastern Plains Metrics
}
