% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys.R
\name{metrics_valleys}
\alias{metrics_valleys}
\title{Valleys-All metrics}
\usage{
metrics_valleys(master, Long, ffg_col)
}
\arguments{
\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{Info}{= Taxonomic Information}

\item{Family}{= Familial taxonomic counts}

\item{Order}{= Ordinal taxonomic counts}

\item{Long_Data}{= Taxonomic data in long format}

\item{Level}{= Taxonomic level ("FAMILY" or "GENUS")}
}
\value{
2011 Chessie BIBI metrics for the all of the valley ecoregion
}
\description{
Valleys-All metrics
}
