% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys_lv.R
\name{metrics_valleys_lv}
\alias{metrics_valleys_lv}
\title{Valleys_LV metrics}
\usage{
metrics_valleys_lv(master, Long, aspt_col, ffg_col)
}
\arguments{
\item{Long}{= Taxonomic data in long format}

\item{aspt_col}{= Specify ASPT column.}

\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{Info}{= Taxonomic Information}
}
\value{
2011 Chessie BIBI metrics for the all of the valley ecoregion
}
\description{
Valleys_LV metrics
}
