% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys_nvl.R
\name{metrics_valleys_nlv}
\alias{metrics_valleys_nlv}
\title{Valleys-NLV metrics}
\usage{
metrics_valleys_nlv(master, Long, aspt_col, ffg_col)
}
\arguments{
\item{Long}{= Taxonomic data in long format}

\item{aspt_col}{= Specify ASPT column.}

\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{Info}{= Taxonomic Information}
}
\value{
2011 Chessie BIBI metrics for the all of the valley ecoregion
}
\description{
Valleys-NLV metrics
}
