% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napu.R
\name{napu}
\alias{napu}
\title{NAPU}
\usage{
napu(Info, Long, taxa.rank, tol_col, ffg_col, scoring)
}
\arguments{
\item{Info}{= Taxonomic Information}

\item{Long}{= Taxonomic data in long format}

\item{tol_col}{= Specify Tolerance value column.}

\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{scoring}{= If scoring is set to "DISCRETE" the 1-3-5 method will be used.
If scoring is set to "GRADIENT" a continuous scoring procedure will be used.}
}
\value{
Prepare the taxonomic data and score all metrics for each sites
within the NAPU bioregion
}
\description{
NAPU
}
