% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napu.R
\name{napu_clean_taxa}
\alias{napu_clean_taxa}
\title{NAPU: Standarize Taxonomic Resolution}
\usage{
napu_clean_taxa(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Removes and adjusts taxonomic resolution to make data sets from
multiple sources comparable.
}
\description{
NAPU: Standarize Taxonomic Resolution
}
