% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.R
\name{nca_clean_taxa}
\alias{nca_clean_taxa}
\title{NCA: Standarize Taxonomic Resolution}
\usage{
nca_clean_taxa(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Removes and adjusts taxonomic resolution to make data sets from
multiple sources comparable.
}
\description{
NCA: Standarize Taxonomic Resolution
}
