% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_and_done.R
\name{one_and_done}
\alias{one_and_done}
\title{One and done}
\usage{
one_and_done(Info, Master, Taxa, WQ, Habitat, TAB_EVENT, TAB_STATIONS,
  TAB_PROJECT)
}
\arguments{
\item{Info}{= Taxonomic information}

\item{Master}{= Taxa count data}

\item{Taxa}{= Taxonomic counts in long data format}

\item{WQ}{= Water quality data in a long data format}

\item{Habitat}{= Habitat data in a long data format}

\item{TAB_EVENT}{= EVENT_ID}

\item{TAB_STATIONS}{= Station information (STATION_ID, ECOREGION, STRAHLER)}

\item{TAB_PROJECT}{= Project information}

\item{Taxonomy}{= TSN #'s with necessary taxonomic levels (CLASS, ORDER, FAMILY)}
}
\value{
Merge taxa water quality, and habitat data into one large data frame.
}
\description{
One and done
}
