% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_resolution.R
\name{output_agency_taxa_diff2}
\alias{output_agency_taxa_diff2}
\title{Write csv files containing agency/program taxonomic discrepancies}
\usage{
output_agency_taxa_diff2(long, bioregion)
}
\arguments{
\item{long}{= long data frame containing taxonomic counts.}

\item{bioregion}{= The bioregion code used to aggregate the data.}
}
\value{
Creates multiple data frames containing taxa that are observed by >= 1 
agency/program and not observed by >= 1 agency/program at the phylum,
subphylum, class, subclass, and order levels. These data frames
also contains the percentage of data that each agency/program composes
in the bioregion being investigated.
}
\description{
Write csv files containing agency/program taxonomic discrepancies
}
