% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{percentile_summary}
\alias{percentile_summary}
\title{Precentile Summary}
\usage{
percentile_summary(fam.final, vp_thresh, outlier = TRUE)
}
\arguments{
\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\description{
Precentile Summary
}
