% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piedmont.R
\name{pied_clean_taxa}
\alias{pied_clean_taxa}
\title{Piedmont: Standarize Taxonomic Resolution}
\usage{
pied_clean_taxa(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Removes and adjusts taxonomic resolution to make data sets from
multiple sources comparable.
}
\description{
Piedmont: Standarize Taxonomic Resolution
}
