% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{plot.me.old}
\alias{plot.me.old}
\title{Plot Me Old}
\usage{
\method{plot}{me.old}(scores.df, spatial, taxon_rank, todays_date)
}
\arguments{
\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\description{
Plot Me Old
}
