% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_me.R
\name{plot.me}
\alias{plot.me}
\title{Import Output}
\usage{
\method{plot}{me}(scores.df, bioregion)
}
\arguments{
\item{scores.df}{= a data frame containing final index scores.}

\item{bioregion}{= specify the bioregion of interest.}
}
\value{
Box-and-Whisker plots of final index scores aggregated by stream condition class.
}
\description{
Import Output
}
