% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{plot_jack}
\alias{plot_jack}
\title{Plot Jackknife Results}
\usage{
plot_jack(spatial, data_set, calc.date, taxon_rank, todays_date, rmse = TRUE,
  maindir = main.dir, prevdir = previous.subdir)
}
\arguments{
\item{my.data}{= data.frame.}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\description{
Plot Jackknife Results
}
