% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_resolution.R
\name{prep_agency_taxa_res}
\alias{prep_agency_taxa_res}
\title{Prepare data for checking taxonomic resolution discrepancies between agencies/programs}
\usage{
prep_agency_taxa_res(long)
}
\arguments{
\item{long}{= long data frame containing taxonomic counts.}
}
\value{
Concatonates AGENCY_CODE and PROGRAM_CODE, and replaces AGENCY_CODE
in the long data frame with the new concatonation.  This allows the data
to be aggregated by unique a unique agency/program.
}
\description{
Prepare data for checking taxonomic resolution discrepancies between agencies/programs
}
