% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_classification.R
\name{prep_bioregion}
\alias{prep_bioregion}
\title{Add a Bioregion Column}
\usage{
prep_bioregion(Long)
}
\arguments{
\item{Long}{= a long data frame with Ecoregion level 4 classes.}
}
\value{
A column classifying each site into a bioregion based on
the Ecoregion level 4 class.
}
\description{
Add a Bioregion Column
}
