% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_classification.R
\name{prep_class_multi}
\alias{prep_class_multi}
\title{Prepare Class Multi}
\usage{
prep_class_multi(Prep_Env)
}
\arguments{
\item{Prep_ENV}{= the output of the prep_env function.}
}
\value{
A data frame containing environmental data to classify sites.
}
\description{
Prepare Class Multi
}
