% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_ecoregion}
\alias{prep_ecoregion}
\title{Ecoregion and Bioregion Modifications}
\usage{
prep_ecoregion(long)
}
\arguments{
\item{Long}{= Long data frame.}
}
\value{
Assigns Chessie BIBI related ecoregion and bioregion distinctions.
}
\description{
Ecoregion and Bioregion Modifications
}
