% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_classification.R
\name{prep_env}
\alias{prep_env}
\title{Prepare Environmental Parameters}
\usage{
prep_env(Prep.data)
}
\arguments{
\item{Prep.Data}{= the output of the prep_data function.}
}
\value{
A data frame containing environmental parameters of interest,
bioregion classes, and no errors.
}
\description{
Prepare Environmental Parameters
}
