% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_habitat}
\alias{prep_habitat}
\title{Prepare Habitat Data}
\usage{
prep_habitat(habitat.df, agg.sample.num = TRUE, bibi.version)
}
\arguments{
\item{agg.sample.num}{= Should the sample number be treated as seperate
samples (FALSE) or aggregated into a single sample (TRUE)?}

\item{Habitat}{= Habitat data in long format}
}
\value{
Transform habitat data from long to wide format
}
\description{
Prepare Habitat Data
}
