% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq_habitat_bibi_2011.R
\name{prep_habitat_old}
\alias{prep_habitat_old}
\title{Prepare Habitat Data}
\usage{
prep_habitat_old(Habitat, agg_sample_num)
}
\arguments{
\item{Habitat}{= Habitat data in long format}
}
\value{
Transform habitat data from long to wide format
}
\description{
Prepare Habitat Data
}
