% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_river.R
\name{prep_large}
\alias{prep_large}
\title{Large River's PREP}
\usage{
prep_large(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Subset the data to only sites in the Chesapeake Bay  watershed with
a Strahler Stream order >= 5
}
\description{
Large River's PREP
}
