% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macs.R
\name{prep_mac}
\alias{prep_mac}
\title{Middle Atlantic Coastal Plain Data PREP}
\usage{
prep_mac(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Subset the data to only sites in the Middle Atlantic Coastal Plain bioregion
with a Strahler Stream order <= 4 and data collected with a kicknet.
}
\description{
Middle Atlantic Coastal Plain Data PREP
}
