% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_merge}
\alias{prep_merge}
\title{Merge taxa, water quality, and habitat data into one}
\usage{
prep_merge(TAB_EVENT, TAXA_PREP, TAB_STATIONS, TAB_WQ, TAB_HABITAT, TAB_PROJECT,
  TAB_HUC)
}
\arguments{
\item{TAB_EVENT}{= EVENT_ID}

\item{TAXA_PREP}{= Taxa data in a long data format}

\item{TAB_STATIONS}{= Station information (STATION_ID, ECOREGION, STRAHLER)}

\item{TAB_WQ}{= Water quality data in a wide format}

\item{TAB_HABITAT}{= Habitat data in a wide format}

\item{TAB_PROJECT}{= Project information}

\item{TAB_HUC}{= HUC information merged on HUC-12}
}
\value{
Merge taxa water quality, and habitat data into one large data frame.
}
\description{
Merge taxa, water quality, and habitat data into one
}
