% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napu.R
\name{prep_napu}
\alias{prep_napu}
\title{NAPU Data PREP}
\usage{
prep_napu(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Subset the data to only sites in the NAPU bioregion with a
Strahler Stream order <= 4
}
\description{
NAPU Data PREP
}
