% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.R
\name{prep_nca}
\alias{prep_nca}
\title{NCA Data PREP}
\usage{
prep_nca(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Subset the data to only sites in the NCA bioregion with a
Strahler Stream order <= 4.
}
\description{
NCA Data PREP
}
