% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piedmont.R
\name{prep_pied}
\alias{prep_pied}
\title{Piedmont Data PREP}
\usage{
prep_pied(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Subset the data to only sites in the Piedmont bioregion with a
Strahler Stream order <= 4
}
\description{
Piedmont Data PREP
}
