% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_river.R
\name{prep_pied_large}
\alias{prep_pied_large}
\title{Piedmont Large River's PREP}
\usage{
prep_pied_large(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Subset the data to only sites in the Piedmont bioregion with a
Strahler Stream order >= 5
}
\description{
Piedmont Large River's PREP
}
