% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_barplots.R
\name{prep_rate_stacks}
\alias{prep_rate_stacks}
\title{Prep Rate Stacks}
\usage{
prep_rate_stacks(spatial, previous.subdir, calc.date, very.poor.thresh,
  random_event)
}
\arguments{
\item{spatial}{= "BASIN", "REGION", or "BIOREGION".}

\item{previous.subdir}{= specify the location of the previously calculated data.}
}
\value{
Prepares the data for the stacked barplots of the ratings.
}
\description{
Prep Rate Stacks
}
