% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridges.R
\name{prep_ridges}
\alias{prep_ridges}
\title{Ridges Data PREP}
\usage{
prep_ridges(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Subset the data to only sites in the Ridges bioregion with a
Strahler Stream order <= 4
}
\description{
Ridges Data PREP
}
