% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_river.R
\name{prep_ridges_large}
\alias{prep_ridges_large}
\title{Ridges Large River's PREP}
\usage{
prep_ridges_large(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Subset the data to only sites in the Ridges bioregion with a
Strahler Stream order >= 5
}
\description{
Ridges Large River's PREP
}
