% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep.R
\name{prep_sep}
\alias{prep_sep}
\title{Southeastern Plains Data PREP}
\usage{
prep_sep(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Subset the data to only sites in the Southeastern Plains bioregion
with a Strahler Stream order <= 4 and data collected with a kicknet.
}
\description{
Southeastern Plains Data PREP
}
