% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_subset}
\alias{prep_subset}
\title{Prepare subset specific to BIBI}
\usage{
prep_subset(long.df)
}
\arguments{
\item{long.df}{= Data in a long data format.}
}
\value{
A subset of the data containing sampling events with strahler stream
order less than or equal to 4 and samples collected with a kick net or 
another method considered comparable.
}
\description{
Prepare subset specific to BIBI
}
