% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_taxa}
\alias{prep_taxa}
\title{Prepare Taxonomic Counts}
\usage{
prep_taxa(master.df, taxa.df, clean.taxa)
}
\arguments{
\item{Master}{= Taxa count data}

\item{Taxa}{= Taxonomic counts in long data format}
}
\value{
Merges a taxonomic counts data frame with a master taxa list
containing the following information for each taxon when applicable:
TSN, Phylum, Subphylum, Class, Subclass, Order, Suborder, Family,
Subfamily, Tribe, Genus, and Species.
}
\description{
Prepare Taxonomic Counts
}
