% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys.R
\name{prep_valleys}
\alias{prep_valleys}
\title{Valley's Data PREP}
\usage{
prep_valleys(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Subset the data to only sites in the Valley's bioregion with a
Strahler Stream order <= 4
}
\description{
Valley's Data PREP
}
