% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_river.R
\name{prep_valleys_large}
\alias{prep_valleys_large}
\title{Valleys-All Large River's PREP}
\usage{
prep_valleys_large(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Subset the data to only sites in the Valleys bioregion with a
Strahler Stream order >= 5
}
\description{
Valleys-All Large River's PREP
}
