% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys_lv.R
\name{prep_valleys_lv}
\alias{prep_valleys_lv}
\title{Valleys_LV PREP}
\usage{
prep_valleys_lv(Long)
}
\arguments{
\item{Long=}{Taxonomic data in long format}
}
\value{
2011 Chessie BIBI metrics and scores for the ridge ecoregion
}
\description{
Valleys_LV PREP
}
