% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys_nvl.R
\name{prep_valleys_nlv}
\alias{prep_valleys_nlv}
\title{Valleys_NLV PREP}
\usage{
prep_valleys_nlv(Long)
}
\arguments{
\item{Long=}{Taxonomic data in long format}
}
\value{
2011 Chessie BIBI metrics and scores for the ridge ecoregion
}
\description{
Valleys_NLV PREP
}
