% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_wq}
\alias{prep_wq}
\title{Prepare Water Quality Data}
\usage{
prep_wq(tab.wq)
}
\arguments{
\item{WQ}{= Water quality data in long format}
}
\value{
Transform water quality data from long to wide format
}
\description{
Prepare Water Quality Data
}
