% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq_habitat_bibi_2011.R
\name{prep_wq_old}
\alias{prep_wq_old}
\title{Prepare Water Quality Data}
\usage{
prep_wq_old(WQ)
}
\arguments{
\item{WQ}{= Water quality data in long format}

\item{agg_sample_num}{= Should the sample number be treated as seperate
samples (FALSE) or aggregated into a single sample (TRUE)?}
}
\value{
Transform water quality data from long to wide format
}
\description{
Prepare Water Quality Data
}
