% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{rate.me2}
\alias{rate.me2}
\title{Rate Me 2}
\usage{
rate.me2(scores.df, percentiles, taxa.rank = "FAMILY", napu = TRUE)
}
\arguments{
\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\description{
Rate Me 2
}
