% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_barplots.R
\name{rate_stacks}
\alias{rate_stacks}
\title{Rate Stacks}
\usage{
rate_stacks(spatial, previous.subdir, calc.date, todays_date, very.poor.thresh,
  random_event)
}
\arguments{
\item{spatial}{= "BASIN", "REGION", or "BIOREGION".}

\item{previous.subdir}{= specify the location of the previously calculated data.}
}
\value{
Creates stacked barplots of the ratings.
}
\description{
Rate Stacks
}
