% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_me.R
\name{read.me}
\alias{read.me}
\title{Import Output}
\usage{
read.me(high.res.front, high.res.proc_date, high.res.data_set, low.res.front,
  low.res.proc_date, low.res.data_set, job = NULL)
}
\arguments{
\item{high.res.front}{= specify the common prefix to the high resolution bioregion output files.}

\item{high.res.proc_date}{= specify the date associated with the high resolution bioregion output files.}

\item{high.res.data_set}{= specify the data set of interest associated with the high resolution bioregion output files.}

\item{low.res.front}{= specify the common prefix to the low resolution bioregion output files.}

\item{low.res.proc_date}{= specify the date associated with the low resolution bioregion output files.}

\item{low.res.data_set}{= specify the data set of interest associated with the low resolution bioregion output files.}

\item{job}{= When equal to NULL the data is not modified beyond the file imported, just 
aggregated with the other bioregion. When equal to "METRICS" the data is imported, transformed
to a long data type, and aggregated with the other bioregions.}
}
\value{
Import and aggregate BIBI output files.
}
\description{
Import Output
}
