% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ref_sev_percentiles}
\alias{ref_sev_percentiles}
\title{Caclulate Reference and Severely Degraded percentiles}
\usage{
ref_sev_percentiles(all.scores, spatial, taxon.rank, todays.date)
}
\description{
Caclulate Reference and Severely Degraded percentiles
}
