% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_classification.R
\name{remove_errors}
\alias{remove_errors}
\title{Remove Environmental Paramater Errors}
\usage{
remove_errors(env.df)
}
\arguments{
\item{env.df}{= a data frame contatining environmental parameters.}
}
\value{
Removes rows with environmental parameter errors.
}
\description{
Remove Environmental Paramater Errors
}
