% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridges.R
\name{ridge_clean_taxa}
\alias{ridge_clean_taxa}
\title{Ridge: Standarize Taxonomic Resolution}
\usage{
ridge_clean_taxa(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Removes and adjusts taxonomic resolution to make data sets from
multiple sources comparable.
}
\description{
Ridge: Standarize Taxonomic Resolution
}
