% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridges.R
\name{ridges}
\alias{ridges}
\title{Ridges}
\usage{
ridges(Info, Long, ffg_col, habit_col, scoring)
}
\arguments{
\item{Info}{= Taxonomic Information}

\item{Long}{= Taxonomic data in long format}

\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{habit_col}{= Specify Habit column.}

\item{scoring}{= If scoring is set to "DISCRETE" the 1-3-5 method will be used.
If scoring is set to "GRADIENT" a continuous scoring procedure will be used.}
}
\value{
Prepare the taxonomic data and score all metrics for each sites
within the Ridges bioregion
}
\description{
Ridges
}
