% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_1_3_5}
\alias{score_1_3_5}
\title{1-3-5 Scoring}
\usage{
score_1_3_5(score.df, thresh.df, metric, disturbance)
}
\arguments{
\item{score.df}{= a dataframe containing raw metric values.}

\item{thresh.df}{= a matrix containing thresholds for scoring.}

\item{metric}{= a string identifying the metric of interest.}

\item{distrubance}{= A string identifying if the metric increases
("INCREASE") or decreases ("DECREASE") with distrubance.}
}
\value{
Scores metrics on a 1-3-5 scale based on the thresholds from the 2011
Chessie BIBI
}
\description{
1-3-5 Scoring
}
