% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macs.R
\name{score_mac}
\alias{score_mac}
\title{Score Middle Atlantic Coastal Plain Metrics}
\usage{
score_mac(Info, Long, taxa.rank, tol_col, habit_col)
}
\arguments{
\item{Info}{= Taxonomic Information}

\item{Long}{= Taxonomic data in long format}

\item{tol_col}{= Specify Tolerance value column.}

\item{Level}{= Taxonomic level ("FAMILY" or "GENUS")}
}
\value{
Metric scores for sites within the Middle Atlantic Coastal Plain bioregion
}
\description{
Score Middle Atlantic Coastal Plain Metrics
}
