% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napu.R
\name{score_napu}
\alias{score_napu}
\title{Score NAPU Metrics}
\usage{
score_napu(Info, Long, taxa.rank, tol_col, ffg_col, scoring)
}
\arguments{
\item{Info}{= Taxonomic Information}

\item{Long}{= Taxonomic data in long format}

\item{tol_col}{= Specify Tolerance value column.}

\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{scoring}{= If scoring is set to "DISCRETE" the 1-3-5 method will be used.
If scoring is set to "GRADIENT" a continuous scoring procedure will be used.}
}
\value{
Metric scores for sites within the NAPU bioregion
}
\description{
Score NAPU Metrics
}
