% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep.R
\name{score_sep}
\alias{score_sep}
\title{Score Southeastern Plains Metrics}
\usage{
score_sep(Info, Long)
}
\arguments{
\item{Info}{= Taxonomic Information}

\item{Long}{= Taxonomic data in long format}
}
\value{
Metric scores for sites within the Southeastern Plains bioregion
}
\description{
Score Southeastern Plains Metrics
}
