% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys.R
\name{score_valleys}
\alias{score_valleys}
\title{Score valley's Metrics}
\usage{
score_valleys(Info, Long, ffg_col, scoring)
}
\arguments{
\item{Info}{= Taxonomic Information}

\item{Long}{= Taxonomic data in long format}

\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{scoring}{= If scoring is set to "DISCRETE" the 1-3-5 method will be used.
If scoring is set to "GRADIENT" a continuous scoring procedure will be used.}
}
\value{
Metric scores for sites within the Valley's bioregion
}
\description{
Score valley's Metrics
}
