% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys_nvl.R
\name{score_valleys_nlv}
\alias{score_valleys_nlv}
\title{Valleys NVL}
\usage{
score_valleys_nlv(Info, Long, aspt_col, ffg_col, scoring)
}
\arguments{
\item{Info}{= Taxonomic Information}

\item{Long}{= Taxonomic data in long format}

\item{aspt_col}{= Specify ASPT column.}

\item{ffg_col}{= Specify Funtional Feeding Group column.}

\item{scoring}{= If scoring is set to "DISCRETE" the 1-3-5 method will be used.
If scoring is set to "GRADIENT" a continuous scoring procedure will be used.}
}
\value{
2011 Chessie BIBI metrics and scores for the valleys ecoregion
}
\description{
Valleys NVL
}
