% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep.R
\name{sep}
\alias{sep}
\title{Southeastern Plains}
\usage{
sep(Info, Long)
}
\arguments{
\item{Info}{= Taxonomic Information}

\item{Long}{= Taxonomic data in long format}
}
\value{
Prepare the taxonomic data and score all metrics for each sites
within the Southeastern Plains bioregion
}
\description{
Southeastern Plains
}
