% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep.R
\name{sep_clean_taxa}
\alias{sep_clean_taxa}
\title{SEP: Standarize Taxonomic Resolution}
\usage{
sep_clean_taxa(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Removes and adjusts taxonomic resolution to make data sets from
multiple sources comparable.
}
\description{
SEP: Standarize Taxonomic Resolution
}
