% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_classification.R
\name{shrink_env}
\alias{shrink_env}
\title{Just Important Environmental Parameters}
\usage{
shrink_env(Prep.Data)
}
\arguments{
\item{Prep.Data}{= the output of the prep_data function.}
}
\value{
A data frame of environmental parameters of interest.
}
\description{
Just Important Environmental Parameters
}
