% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_classification.R
\name{site_classification}
\alias{site_classification}
\title{Site Classification}
\usage{
site_classification(env.param)
}
\arguments{
\item{env.param}{= a data frame of environmental parameters representitive
of each sampling event.  Needs to include water quality parameters
(Specific Conductivity, pH, and Dissolved Oxygen), habitat parameters
(Rapid Habitat Assessment), and (watershed parameters()).}
}
\value{
Each sampling event is classified as Reference, Mixed, or Degraded
based on environmental parameters. If environmental parameters are missing
the site cannot be accurately classified as Reference or Degraded.
Therefore, the site is classified as Mixed.  The Mixed class does not
influence the selection of metrics but will be scored and assigned an
IBI rating.
}
\description{
Site Classification
}
