% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq_habitat_bibi_2011.R
\name{site_classification_2011}
\alias{site_classification_2011}
\title{Site Classification (2011)}
\usage{
site_classification_2011(env.param)
}
\arguments{
\item{env.parm}{= a data frame containing the necessary habitat and water
quality variables to classify sites based on the 2011 Chessie BIBI
protocol.}
}
\value{
A dataframe containg site classes (CATEGORY).
}
\description{
Site Classification (2011)
}
