% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_resolution.R
\name{taxa_absence}
\alias{taxa_absence}
\title{Taxa Absent from Agency/Program}
\usage{
taxa_absence(long, rank, col.ord = NULL)
}
\arguments{
\item{long}{= long data frame containing taxonomic counts.}

\item{rank}{= Taxonomic rank used to aggregate the data.}

\item{col.ord}{= a list of names used to sort the final output.}
}
\value{
Returns a data frame containing taxa that are observed by >= 1 
agency/program and not observed by >= 1 agency/program.
}
\description{
Taxa Absent from Agency/Program
}
