% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_resolution.R
\name{test_id}
\alias{test_id}
\title{Test Agency/Program Standard Taxonomic Resolution of Specific Taxa}
\usage{
test_id(long, low.rank, upper.rank, taxon, bioregion)
}
\arguments{
\item{long}{= long data frame containing taxonomic counts.}

\item{low.rank}{= The lower of taxonomic rank of interest.}

\item{upper.rank}{= The upper of taxonomic rank of interest; corresponds with
the taxon of interest.}

\item{taxon}{= The taxon of interest.}

\item{bioregion}{= The bioregion code used to aggregate the data.}
}
\value{
Returns a list of names at the taxonomic level specified by low.rank,
aggregated by the taxon of interest. The data frame
also contains the percentage of data that each agency/program composes
in the bioregion being investigated.
}
\description{
Test Agency/Program Standard Taxonomic Resolution of Specific Taxa
}
