% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys.R
\name{val_clean_taxa}
\alias{val_clean_taxa}
\title{Valleys: Standarize Taxonomic Resolution}
\usage{
val_clean_taxa(Long)
}
\arguments{
\item{Long}{= Taxonomic data in long format}
}
\value{
Removes and adjusts taxonomic resolution to make data sets from
multiple sources comparable.
}
\description{
Valleys: Standarize Taxonomic Resolution
}
