% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvsci.R
\name{wvsci_metrics}
\alias{wvsci_metrics}
\title{Calculate the metrics from WVSCI}
\usage{
wvsci_metrics(Long)
}
\arguments{
\item{Long}{= a long data frame of taxonomic data}
}
\value{
Calculates the metrics from the West Virginia Stream Condition Index
(WVSCI) (Tetra Tech, Inc. 2000).
}
\description{
Calculate the metrics from WVSCI
}
