% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_tables_figures.R
\name{year_counts_barplot}
\alias{year_counts_barplot}
\title{Year Counts}
\usage{
year_counts_barplot(spatial, taxon.rank, todays.date)
}
\arguments{
\item{spatial}{= "BASIN", "REGION", or "BIOREGION".}

\item{taxon.rank}{= "ORDER", "FAMILY", or "GENUS".}

\item{todays.date}{= todays date.}
}
\value{
Provides a a bar plot figure of sample counts by year.
}
\description{
Year Counts
}
