% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{clean_taxa}
\alias{clean_taxa}
\title{Clean Taxa}
\usage{
clean_taxa(long)
}
\arguments{
\item{Bioregion.df}{= Bioregion data frame to be aggregated}
}
\value{
Standardizes taxa according to the Chessie BIBI standards.
}
\description{
Clean Taxa
}
