% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_me.R, R/simple_tables_figures.R
\name{plot.me}
\alias{plot.me}
\alias{plot.me}
\title{Import Output}
\usage{
\method{plot}{me}(scores.df, spatial, taxon_rank, todays_date)

\method{plot}{me}(scores.df, spatial, taxon_rank, todays_date)
}
\arguments{
\item{scores.df}{= a data frame containing final index scores.}

\item{bioregion}{= specify the bioregion of interest.}

\item{my.data}{=}

\item{index_res}{= Spatial Resolution (i.e., "BASIN", "COAST", "INLAND", or
"BIOREGION")}

\item{bioregion}{= If index_res is set to "BIOREGION," then a list of 
bioregions to keep must be specified.}
}
\value{
Box-and-Whisker plots of final index scores aggregated by stream condition class.
}
\description{
Import Output

Plot Me
}
