% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_subset}
\alias{prep_subset}
\title{Prepare subset specific to BIBI}
\usage{
prep_subset(Long)
}
\arguments{
\item{Long}{= Data in a long format.}
}
\value{
A subset of the data containing sampling events with strahler stream
order less than or equal to 4 and methods comparable to kick net samples.
}
\description{
Prepare subset specific to BIBI
}
